+++ src/main/java/oshi/software/os/windows/nt/GlobalMemory.java
@@ -26,12 +26,10 @@ public class GlobalMemory implements Memory {
     	}
 	}
 	
-	@Override
 	public long getAvailable() {
 		return _memory.ullAvailPhys.longValue();
 	}
 
-	@Override
 	public long getTotal() {
 		return _memory.ullTotalPhys.longValue();
 	}